Fixed compiler warning

+++ src/ptlib/unix/tlibthrd.cxx
@@ -818,7 +818,7 @@ PBoolean PThread::IsTerminated() const
      So, if we have a /proc, we always use that mechanism to determine
      if the external thread still exists. */
   char fn[100];
-  snprintf(fn, sizeof(fn), "/proc/%u/task/%u/stat", getpid(), GetUniqueIdentifier());
+  snprintf(fn, sizeof(fn), "/proc/%u/task/%u/stat", getpid(), (unsigned)(intptr_t)GetUniqueIdentifier());
   if (access(fn, R_OK) == 0)
     return false;
